--- tests/apiserver-proxy-tests/apiserver-proxy/internal/client.go
@@ -1,18 +1,0 @@
-package graphql
-
-import (
-	"crypto/tls"
-	"net/http"
-)
-
-func Authenticate(config idProviderConfig) (string, error) {
-	httpClient := &http.Client{
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
-		},
-	}
-
-	idTokenProvider := newDexIdTokenProvider(httpClient, config)
-	token, err := idTokenProvider.fetchIdToken()
-	return token, err
-}
