+++ tests/apiserver-proxy-tests/fetch-token/internal/auth.go
@@ -1,0 +1,18 @@
+package internal
+
+import (
+	"crypto/tls"
+	"net/http"
+)
+
+func Authenticate(config idProviderConfig) (string, error) {
+	httpClient := &http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: &tls.Config{InsecureSkipVerify: true},
+		},
+	}
+
+	idTokenProvider := newDexIdTokenProvider(httpClient, config)
+	token, err := idTokenProvider.fetchIdToken()
+	return token, err
+}
